- Added connection failure testing - Changed delete and back function in DisplayECG - Added sample files to ECG Test

+++ Heart_Application_Monitor/app/src/main/java/edu/und/cs/com/heart_monitor/ECG.java
@@ -1,8 +1,10 @@
 package edu.und.cs.com.heart_monitor;
 
+import android.app.AlertDialog;
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
 import android.bluetooth.BluetoothSocket;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.AsyncTask;
@@ -32,10 +34,10 @@ public class ECG extends RoboActivity implements View.OnClickListener {
     boolean runTest = false;
     boolean testFailed = false;
     boolean connectionFailure = false;
+    boolean isCanceled = false;
     private static final UUID MY_UUID = UUID.fromString("00001101-0000-1000-8000-00805F9B34FB");
     private boolean testInitiated = false;
     private AsyncTask myThread;
-    private int RSSI;                                           //bluetooth signal strength TODO ANDREW
     public FileHelper myFileHelper;
     private BluetoothSocket sock = null;
     private BITalinoDevice bitalino;
@@ -47,7 +49,7 @@ public class ECG extends RoboActivity implements View.OnClickListener {
         setContentView(R.layout.fragment_ecg);
 
         //let user know it will take a few seconds to start getting readings
-        makeText(this, "Establishing Connection to Sensor", Toast.LENGTH_LONG).show();
+        Toast.makeText(this, "Press start to begin ECG", Toast.LENGTH_LONG).show();
 
         myFileHelper = new FileHelper();
 
@@ -69,30 +71,9 @@ public class ECG extends RoboActivity implements View.OnClickListener {
     public void onClick(View v) {
         switch (v.getId()) {
             case R.id.startBTN:
-                if (!testInitiated) {
-                    runTest = true;
-                    myFileHelper.startFile(myFileHelper, getApplicationContext(), false);
-                    myThread = new TestAsyncTask().execute();
-                } else {
-                    //close everything and delete the file
-                    try {
-                        bitalino.stop();         //signal board to quit sending packets
-                        sock.close();               //close socket on this end
-                    } catch (Exception e) {
-                        Log.e(TAG, "There was an error.", e);
-                    }
-                    myThread.cancel(true);
-                    runTest = false;                                                                                    //ensure current test terminates correctly
-                    myFileHelper.closeFile(myFileHelper, getApplicationContext());                                       //close file output stream
-                    getApplicationContext().deleteFile(myFileHelper.fileName);
-
-                    try {   //make sure socket has enough time to close
-                        Thread.sleep(5000);
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-
-                    //restart the process
+                if (!runTest && !testInitiated) {
+                    //let user know it will take a few seconds to start getting readings
+                    makeText(this, "Establishing Connection to Sensor", Toast.LENGTH_LONG).show();
                     runTest = true;
                     myFileHelper.startFile(myFileHelper, getApplicationContext(), false);
                     myThread = new TestAsyncTask().execute();
@@ -100,6 +81,7 @@ public class ECG extends RoboActivity implements View.OnClickListener {
                 break;
             case R.id.quitBTN:
                 if ((runTest) && (!testFailed) && (testInitiated)) {                       //test has run successfully, has stopped, can store file
+                    isCanceled = true;
                     try {
                         bitalino.stop();         //signal board to quit sending packets
                         sock.close();               //close socket on this end
@@ -115,14 +97,24 @@ public class ECG extends RoboActivity implements View.OnClickListener {
                     newIntent.putExtras(myBundle);
                     startActivity(newIntent);
                 } else if ((!runTest) && (testFailed) && (!testInitiated)) {                       //test is not running and an error of some kind occurred
+                    isCanceled = true;
                     Toast.makeText(this, "No recording made", Toast.LENGTH_LONG).show();
                 }
                 break;
             case R.id.cancelBTN:
+                isCanceled = true;
                 runTest = false;
-                myFileHelper.closeFile(myFileHelper, getApplicationContext());
-                getApplicationContext().deleteFile(myFileHelper.fileName);
-                android.os.Process.killProcess(android.os.Process.myPid());
+                if (myThread != null) {
+                    try {
+                        bitalino.stop();         //signal board to quit sending packets
+                        sock.close();               //close socket on this end
+                    } catch (Exception e) {
+                        Log.e(TAG, "There was an error.", e);
+                    }
+                    myThread.cancel(true);
+                    myFileHelper.closeFile(myFileHelper, getApplicationContext());
+                    getApplicationContext().deleteFile(myFileHelper.fileName);
+                }
                 startActivity(new Intent(ECG.this, MainActivity.class));
                 break;
         }
@@ -133,6 +125,26 @@ public class ECG extends RoboActivity implements View.OnClickListener {
         // Inflate the menu; this adds items to the action bar if it is present.
         // getMenuInflater().inflate(R.menu.main, menu);
         return true;
+    }
+
+    private void handleConnectionFailure() {
+        if (!isCanceled) {
+            myThread.cancel(true);
+            AlertDialog.Builder builder = new AlertDialog.Builder(ECG.this);
+            builder.setMessage("Error Connecting! Check your connection in User Settings.")
+                    .setCancelable(false)
+                    .setPositiveButton("OK", new DialogInterface.OnClickListener() {
+                        public void onClick(DialogInterface dialog, int id) {
+                            if (myFileHelper.fileName != null) {
+                                myFileHelper.closeFile(myFileHelper, getApplicationContext());
+                                getApplicationContext().deleteFile(myFileHelper.fileName);
+                            }
+                            startActivity(new Intent(ECG.this, MainActivity.class));
+                        }
+                    });
+            AlertDialog alert = builder.create();
+            alert.show();
+        }
     }
 
     private class TestAsyncTask extends AsyncTask<Void, String, Void> {
@@ -162,7 +174,7 @@ public class ECG extends RoboActivity implements View.OnClickListener {
                     bitalino.start();
                     ((GifDrawable) loadingImage.getDrawable()).start();
                     // read until task is stopped
-                    int counter = 0;
+                    currentFrameNumber = 0;
                     while (runTest) {
                         BITalinoFrame[] frames = bitalino.read(sampleNumberToGet);   //read(number of frames to read)
                         if (UPLOAD) {
@@ -176,10 +188,9 @@ public class ECG extends RoboActivity implements View.OnClickListener {
                             //analog 2 == ECG
                             currentValue = frame.getAnalog(ecgChannel);
 
-                            currentFrameNumber = counter;
                             //output results to screen using onProgressUpdate()
-                            publishProgress(Integer.toString(currentValue), Integer.toString(counter));
-                            counter++;
+                            publishProgress(Integer.toString(currentValue), Integer.toString(currentFrameNumber));
+                            currentFrameNumber++;
                         }
                     }
                 } catch (Exception e) {              //error opening socket
@@ -205,25 +216,11 @@ public class ECG extends RoboActivity implements View.OnClickListener {
         protected void onProgressUpdate(String... values) {
             //If the connection has failed, show a message
             if (connectionFailure) {
-                Toast.makeText(getApplicationContext(), "Unable to establish connection", Toast.LENGTH_LONG).show();
+                handleConnectionFailure();
+                //Toast.makeText(getApplicationContext(), "Unable to establish connection", Toast.LENGTH_LONG).show();
             } else {
-                myFileHelper.appendFile(myFileHelper, currentFrameNumber / sampleNumberToGet, currentValue, getApplicationContext());
+                myFileHelper.appendFile(myFileHelper, Integer.parseInt(values[1]) / sampleNumberToGet, Integer.parseInt(values[0]), getApplicationContext());
             }
         }
-
-        @Override
-        protected void onCancelled() {
-            // stop acquisition and close bluetooth connection
-            try {
-                bitalino.stop();         //signal board to quit sending packets
-                //InputStream is = null;      //close input and output streams
-                //OutputStream os = null;
-                sock.close();               //close socket on this end
-            } catch (Exception e) {
-                Log.e(TAG, "There was an error.", e);
-            }
-        }
-
     }
-
 }
